Minor cleanup, delete policy_graident, sync previous Policy gradient related script to pg.py
