+++ README.md
@@ -9,6 +9,6 @@ Init submodules:
 ```
 2. Model Detail
 
-recurrent_mlsh_v1: weighted average of sub-policies
-recurrent_mlsh_v2: choose one sub-policy strictly
-recurrent_mlsh_v3: allow multi-cell RNN and LSTM
+* recurrent_mlsh_v1: weighted average of sub-policies
+* recurrent_mlsh_v2: choose one sub-policy strictly
+* recurrent_mlsh_v3: allow multi-cell RNN and LSTM
