+++ pg.py
@@ -14,7 +14,8 @@ from utils.general import export_plot, get_logger
 
 class PolicyGradient(object):
     def save_model_checkpoint(self, session, saver, filename, epoch_num):
-        save_path = saver.save(session, os.path.expanduser(filename), epoch_num)
+        save_path = saver.save(session, os.path.expanduser(filename),
+                               epoch_num)
         print("\nCheckpoint saved in file: %s" % save_path)
 
     def recover_model_checkpoint(self, session, saver, checkpoint_path):
@@ -22,23 +23,28 @@ class PolicyGradient(object):
         saver.restore(session, checkpoint_path)
         print("Model restored!\n")
 
-    def policy_network(self, mlp_input, output_size, scope,
+    def policy_network(self,
+                       mlp_input,
+                       output_size,
+                       scope,
                        size=config.baseline_layer_size,
-                       n_layers=config.n_layers, output_activation=None):
+                       n_layers=config.n_layers,
+                       output_activation=None):
         out = mlp_input
         with tf.variable_scope(scope):
             for i in range(n_layers):
-                out = layers.fully_connected(out, size,
-                                             activation_fn=tf.nn.relu,
-                                             reuse=False)
+                out = layers.fully_connected(
+                    out, size, activation_fn=tf.nn.relu, reuse=False)
 
-            out = layers.fully_connected(out, output_size,
-                                         activation_fn=output_activation,
-                                         reuse=False)
+            out = layers.fully_connected(
+                out, output_size, activation_fn=output_activation, reuse=False)
 
         return out
 
-    def baseline_network(self, mlp_input, output_size, scope,
+    def baseline_network(self,
+                         mlp_input,
+                         output_size,
+                         scope,
                          n_layers=config.n_layers,
                          size=config.baseline_layer_size,
                          output_activation=None):
@@ -46,13 +52,11 @@ class PolicyGradient(object):
         out = mlp_input
         with tf.variable_scope(scope):
             for i in range(n_layers):
-                out = layers.fully_connected(out, size,
-                                             activation_fn=tf.nn.relu,
-                                             reuse=False)
+                out = layers.fully_connected(
+                    out, size, activation_fn=tf.nn.relu, reuse=False)
 
-            out = layers.fully_connected(out, output_size,
-                                         activation_fn=output_activation,
-                                         reuse=False)
+            out = layers.fully_connected(
+                out, output_size, activation_fn=output_activation, reuse=False)
 
         return out
 
@@ -84,13 +88,13 @@ class PolicyGradient(object):
         self.build()
 
     def add_placeholders_op(self):
-        self.observation_placeholder = tf.placeholder(tf.float32, shape=[None,
-                                                                         self.observation_dim])
+        self.observation_placeholder = tf.placeholder(
+            tf.float32, shape=[None, self.observation_dim])
         if self.discrete:
             self.action_placeholder = tf.placeholder(tf.int64, shape=None)
         else:
-            self.action_placeholder = tf.placeholder(tf.float32, shape=[None,
-                                                                        self.action_dim])
+            self.action_placeholder = tf.placeholder(
+                tf.float32, shape=[None, self.action_dim])
 
         # Define a placeholder for advantages
         self.advantage_placeholder = tf.placeholder(tf.float32, shape=None)
@@ -104,13 +108,13 @@ class PolicyGradient(object):
             self.logprob = -tf.nn.sparse_softmax_cross_entropy_with_logits(
                 labels=self.action_placeholder, logits=self.action_logits)
         else:
-            action_means = self.policy_network(self.observation_placeholder,
-                                               self.action_dim, scope=scope)
-            log_std = tf.get_variable('log_std', shape=[self.action_dim],
-                                      trainable=True)
+            action_means = self.policy_network(
+                self.observation_placeholder, self.action_dim, scope=scope)
+            log_std = tf.get_variable(
+                'log_std', shape=[self.action_dim], trainable=True)
             action_std = tf.exp(log_std)
-            multivariate = tfd.MultivariateNormalDiag(loc=action_means,
-                                                      scale_diag=action_std)
+            multivariate = tfd.MultivariateNormalDiag(
+                loc=action_means, scale_diag=action_std)
             self.sampled_action = tf.random_normal(
                 [self.action_dim]) * action_std + action_means
             self.logprob = multivariate.log_prob(self.action_placeholder)
@@ -124,13 +128,15 @@ class PolicyGradient(object):
 
     def add_baseline_op(self, scope="baseline"):
         self.baseline = tf.squeeze(
-            self.baseline_network(self.observation_placeholder, 1, scope=scope))
-        self.baseline_target_placeholder = tf.placeholder(tf.float32,
-                                                          shape=None)
+            self.baseline_network(
+                self.observation_placeholder, 1, scope=scope))
+        self.baseline_target_placeholder = tf.placeholder(
+            tf.float32, shape=None)
         self.baseline_loss = tf.losses.mean_squared_error(
             self.baseline_target_placeholder, self.baseline)
         self.baseline_opt = tf.train.AdamOptimizer(learning_rate=self.lr)
-        self.update_baseline_op = self.baseline_opt.minimize(self.baseline_loss)
+        self.update_baseline_op = self.baseline_opt.minimize(
+            self.baseline_loss)
 
     def build(self):
         self.add_placeholders_op()
@@ -154,15 +160,15 @@ class PolicyGradient(object):
         self.sess.run(init)
 
     def add_summary(self):
-        self.avg_reward_placeholder = tf.placeholder(tf.float32, shape=(),
-                                                     name="avg_reward")
-        self.max_reward_placeholder = tf.placeholder(tf.float32, shape=(),
-                                                     name="max_reward")
-        self.std_reward_placeholder = tf.placeholder(tf.float32, shape=(),
-                                                     name="std_reward")
+        self.avg_reward_placeholder = tf.placeholder(
+            tf.float32, shape=(), name="avg_reward")
+        self.max_reward_placeholder = tf.placeholder(
+            tf.float32, shape=(), name="max_reward")
+        self.std_reward_placeholder = tf.placeholder(
+            tf.float32, shape=(), name="std_reward")
 
-        self.eval_reward_placeholder = tf.placeholder(tf.float32, shape=(),
-                                                      name="eval_reward")
+        self.eval_reward_placeholder = tf.placeholder(
+            tf.float32, shape=(), name="eval_reward")
 
         tf.summary.scalar("Avg_Reward", self.avg_reward_placeholder)
         tf.summary.scalar("Max_Reward", self.max_reward_placeholder)
@@ -219,12 +225,14 @@ class PolicyGradient(object):
             for step in range(self.config.max_ep_len):
                 states.append(state)
 
-                if str(config.env_name).startswith("Fourrooms"):
+                if config.examine_master and str(
+                        config.env_name).startswith("Fourrooms"):
                     room = self.get_room_by_state(state)
                     rooms.append(room)
 
                     chosen_sub_policy, action = self.sess.run(
-                        [self.chosen_index, self.sampled_action], feed_dict={
+                        [self.chosen_index, self.sampled_action],
+                        feed_dict={
                             self.observation_placeholder: [[states[-1]]]
                         })
                     action = action[0]
@@ -232,13 +240,21 @@ class PolicyGradient(object):
                     if room not in rooms_and_sub_policies:
                         rooms_and_sub_policies[room] = []
                     rooms_and_sub_policies[room].append(chosen_sub_policy)
+                elif config.env_name == "Fourrooms-v1":
+                    action = self.sess.run(
+                        self.sampled_action,
+                        feed_dict={
+                            self.observation_placeholder: [[states[-1]]]
+                        })[0]
                 else:
-                    action = self.sess.run(self.sampled_action, feed_dict={
-                        self.observation_placeholder: states[-1][None]
-                    })[0]
+                    action = self.sess.run(
+                        self.sampled_action,
+                        feed_dict={
+                            self.observation_placeholder: states[-1][None]
+                        })[0]
 
-                action = self.epsilon_greedy(action=action,
-                                             eps=self.get_epsilon(t))
+                action = self.epsilon_greedy(
+                    action=action, eps=self.get_epsilon(t))
                 if self.config.render:
                     env.render()
 
@@ -253,16 +269,17 @@ class PolicyGradient(object):
                 if (not num_episodes) and t == self.config.batch_size:
                     break
 
-            if str(config.env_name).startswith(
-                "Fourrooms") and self.config.render:
-                print(
-                    [(states[room], rooms[room]) for room in range(len(rooms))])
+            if config.examine_master and str(config.env_name).startswith(
+                    "Fourrooms") and self.config.render:
+                print([(states[room], rooms[room])
+                       for room in range(len(rooms))])
                 print(Counter(rooms))
                 print(sorted(Counter(rooms), key=lambda i: i[1]))
                 exit()
 
             path = {
-                "observation": np.array(states), "reward": np.array(rewards),
+                "observation": np.array(states),
+                "reward": np.array(rewards),
                 "action": np.array(actions)
             }
             paths.append(path)
@@ -270,7 +287,8 @@ class PolicyGradient(object):
             if num_episodes and episode >= num_episodes:
                 break
 
-        if str(config.env_name).startswith("Fourrooms"):
+        if config.examine_master and str(
+                config.env_name).startswith("Fourrooms"):
             counter_by_room = {}
             for room in rooms_and_sub_policies:
                 counter = Counter(rooms_and_sub_policies[room])
@@ -303,9 +321,11 @@ class PolicyGradient(object):
     def calculate_advantage(self, returns, observations):
         adv = returns
         if self.config.use_baseline:
-            baseline = self.sess.run(self.baseline, feed_dict={
-                self.observation_placeholder: observations
-            })
+            baseline = self.sess.run(
+                self.baseline,
+                feed_dict={
+                    self.observation_placeholder: observations
+                })
             adv = returns - baseline
 
         if self.config.normalize_advantage:
@@ -314,15 +334,17 @@ class PolicyGradient(object):
         return adv
 
     def update_baseline(self, returns, observations):
-        self.sess.run(self.update_baseline_op, feed_dict={
-            self.observation_placeholder: observations,
-            self.baseline_target_placeholder: returns
-        })
+        self.sess.run(
+            self.update_baseline_op,
+            feed_dict={
+                self.observation_placeholder: observations,
+                self.baseline_target_placeholder: returns
+            })
 
     def get_epsilon(self, t):
         return max(config.min_epsilon,
-                   config.max_epsilon - float(t) / config.num_batches * (
-                       config.max_epsilon - config.min_epsilon))
+                   config.max_epsilon - float(t) / config.num_batches *
+                   (config.max_epsilon - config.min_epsilon))
 
     def train(self):
         last_record = 0
@@ -330,11 +352,13 @@ class PolicyGradient(object):
         self.init_averages()
         scores_eval = []
         self.plot = {
-            'room' + str(i): {j: [] for j in range(config.num_sub_policies)} for
-            i in range(4)}
+            'room' + str(i): {j: []
+                              for j in range(config.num_sub_policies)}
+            for i in range(4)
+        }
 
         for t in range(self.config.num_batches):
-            print(t, self.get_epsilon(t))
+            # print(t, self.get_epsilon(t))
             paths, total_rewards = self.sample_path(env=self.env)
 
             scores_eval += total_rewards
@@ -354,11 +378,13 @@ class PolicyGradient(object):
 
             if self.config.use_baseline:
                 self.update_baseline(returns, observations)
-            self.sess.run(self.train_op, feed_dict={
-                self.observation_placeholder: observations,
-                self.action_placeholder: actions,
-                self.advantage_placeholder: advantages
-            })
+            self.sess.run(
+                self.train_op,
+                feed_dict={
+                    self.observation_placeholder: observations,
+                    self.action_placeholder: actions,
+                    self.advantage_placeholder: advantages
+                })
 
             if t % self.config.summary_freq == 0:
                 self.update_averages(total_rewards, scores_eval)
@@ -367,8 +393,8 @@ class PolicyGradient(object):
 
             avg_reward = np.mean(total_rewards)
             sigma_reward = np.sqrt(np.var(total_rewards) / len(total_rewards))
-            msg = "Average reward: {:04.2f} +/- {:04.2f}".format(avg_reward,
-                                                                 sigma_reward)
+            msg = "Average reward: {:04.2f} +/- {:04.2f}".format(
+                avg_reward, sigma_reward)
             self.logger.info(msg)
 
             last_record += 1
@@ -379,29 +405,36 @@ class PolicyGradient(object):
 
             if t % config.record_freq == 0:
                 self.save_model_checkpoint(self.sess, self.saver,
-                                           os.path.join(self.config.output_path,
-                                                        'model.ckpt'), t)
+                                           os.path.join(
+                                               self.config.output_path,
+                                               'model.ckpt'), t)
 
         self.logger.info("- Training done.")
         export_plot(scores_eval, "Score", config.env_name,
                     self.config.plot_output)
 
         if str(config.env_name).startswith(
-            "Fourrooms") and config.examine_master:
+                "Fourrooms") and config.examine_master:
             import matplotlib.pyplot as plt
             plt.rcParams["figure.figsize"] = [12, 12]
-            f, ((ax1, ax2), (ax3, ax4)) = plt.subplots(2, 2, sharex='col',
-                                                       sharey='row')
+            f, ((ax1, ax2), (ax3, ax4)) = plt.subplots(
+                2, 2, sharex='col', sharey='row')
             axes = {'room0': ax1, 'room1': ax2, 'room2': ax3, 'room3': ax4}
             for room in self.plot:
                 axes[room].set_title(room, size=20)
                 for sub in range(config.num_sub_policies):
                     prob_list = self.plot[room][sub]
-                    axes[room].plot(range(len(prob_list)), prob_list,
-                                    linewidth=5)
-                axes[room].legend(['subpolicy' + str(sub) for sub in
-                                   range(config.num_sub_policies)],
-                                  loc='upper left', prop={'size': 20})
+                    axes[room].plot(
+                        range(len(prob_list)), prob_list, linewidth=5)
+                axes[room].legend(
+                    [
+                        'subpolicy' + str(sub)
+                        for sub in range(config.num_sub_policies)
+                    ],
+                    loc='upper left',
+                    prop={
+                        'size': 20
+                    })
             plt.tight_layout()
             plt.savefig('Rooms and Subs', dpi=300)
 
@@ -425,15 +458,18 @@ class PolicyGradient(object):
         paths, rewards = self.sample_path(env, num_episodes)
         avg_reward = np.mean(rewards)
         sigma_reward = np.sqrt(np.var(rewards) / len(rewards))
-        msg = "Average reward: {:04.2f} +/- {:04.2f}".format(avg_reward,
-                                                             sigma_reward)
+        msg = "Average reward: {:04.2f} +/- {:04.2f}".format(
+            avg_reward, sigma_reward)
         self.logger.info(msg)
         return avg_reward
 
     def record(self):
         env = gym.make(self.config.env_name)
-        env = gym.wrappers.Monitor(env, self.config.record_path,
-                                   video_callable=lambda x: True, resume=True)
+        env = gym.wrappers.Monitor(
+            env,
+            self.config.record_path,
+            video_callable=lambda x: True,
+            resume=True)
         self.evaluate(env, 1)
 
     def run(self):
