+++ test_env/__init__.py
@@ -1,5 +1,6 @@
 from gym.envs.registration import register
 
+import gym
 from test_env.envs import *
 
 register(
@@ -9,4 +10,7 @@ register(
         'map_name': '9x9',
     })
 
-register(id='KeyDoor-v1', entry_point='test_env.envs.key_door:KeyDoor', )
+register(
+    id='KeyDoor-v1',
+    entry_point='test_env.envs.key_door:KeyDoor',
+)
